fips SIMD core, bug fixes

+++ src/kernel/analyze-impl.h
@@ -187,10 +187,10 @@ FASTENT_FN(simd_body_impl)(fastent_chunk_state * st,
   __m256i lhs_sad          = _mm256_setzero_si256();
 
   /*  4 histogram banks; bytes 0..3 within each quad hit banks 0..3.  */
-  u32 * FASTENT_RESTRICT b0 = st->bank[0];
-  u32 * FASTENT_RESTRICT b1 = st->bank[1];
-  u32 * FASTENT_RESTRICT b2 = st->bank[2];
-  u32 * FASTENT_RESTRICT b3 = st->bank[3];
+  u64 * FASTENT_RESTRICT b0 = st->bank[0];
+  u64 * FASTENT_RESTRICT b1 = st->bank[1];
+  u64 * FASTENT_RESTRICT b2 = st->bank[2];
+  u64 * FASTENT_RESTRICT b3 = st->bank[3];
 
   /*  MC Pi state hoisted into locals for register residency.  */
   i32 mc_pos     = st->mc_pos;
@@ -439,10 +439,10 @@ FASTENT_FN(simd_body_impl)(fastent_chunk_state * st,
   __m512i scc_acc64      = _mm512_setzero_si512(); /*  8 i64 lanes  */
   __m512i lhs_sad        = _mm512_setzero_si512();
 
-  u32 * FASTENT_RESTRICT b0 = st->bank[0];
-  u32 * FASTENT_RESTRICT b1 = st->bank[1];
-  u32 * FASTENT_RESTRICT b2 = st->bank[2];
-  u32 * FASTENT_RESTRICT b3 = st->bank[3];
+  u64 * FASTENT_RESTRICT b0 = st->bank[0];
+  u64 * FASTENT_RESTRICT b1 = st->bank[1];
+  u64 * FASTENT_RESTRICT b2 = st->bank[2];
+  u64 * FASTENT_RESTRICT b3 = st->bank[3];
 
   i32 mc_pos     = st->mc_pos;
   u8  m0 = st->mc_buf[0], m1 = st->mc_buf[1], m2 = st->mc_buf[2];
@@ -704,10 +704,10 @@ FASTENT_FN(simd_body_impl)(fastent_chunk_state * st,
   __m128i scc_acc64        = _mm_setzero_si128();  /*  2 i64 lanes  */
   __m128i lhs_sad          = _mm_setzero_si128();
 
-  u32 * FASTENT_RESTRICT b0 = st->bank[0];
-  u32 * FASTENT_RESTRICT b1 = st->bank[1];
-  u32 * FASTENT_RESTRICT b2 = st->bank[2];
-  u32 * FASTENT_RESTRICT b3 = st->bank[3];
+  u64 * FASTENT_RESTRICT b0 = st->bank[0];
+  u64 * FASTENT_RESTRICT b1 = st->bank[1];
+  u64 * FASTENT_RESTRICT b2 = st->bank[2];
+  u64 * FASTENT_RESTRICT b3 = st->bank[3];
 
   /*  MC Pi state hoisted into locals.  */
   i32 mc_pos     = st->mc_pos;
@@ -902,10 +902,10 @@ FASTENT_FN(simd_body_impl)(fastent_chunk_state * st,
   int64x2_t  scc_acc64 = vdupq_n_s64(0);
   uint64x2_t lhs_sad   = vdupq_n_u64(0);
 
-  u32 * FASTENT_RESTRICT b0 = st->bank[0];
-  u32 * FASTENT_RESTRICT b1 = st->bank[1];
-  u32 * FASTENT_RESTRICT b2 = st->bank[2];
-  u32 * FASTENT_RESTRICT b3 = st->bank[3];
+  u64 * FASTENT_RESTRICT b0 = st->bank[0];
+  u64 * FASTENT_RESTRICT b1 = st->bank[1];
+  u64 * FASTENT_RESTRICT b2 = st->bank[2];
+  u64 * FASTENT_RESTRICT b3 = st->bank[3];
 
   i32 mc_pos     = st->mc_pos;
   u8  m0 = st->mc_buf[0], m1 = st->mc_buf[1], m2 = st->mc_buf[2];
@@ -1097,10 +1097,10 @@ FASTENT_FN(simd_body_impl)(fastent_chunk_state * st,
   v128_t scc_acc64 = wasm_i64x2_splat(0);
   v128_t lhs_sad   = wasm_i64x2_splat(0);
 
-  u32 * FASTENT_RESTRICT b0 = st->bank[0];
-  u32 * FASTENT_RESTRICT b1 = st->bank[1];
-  u32 * FASTENT_RESTRICT b2 = st->bank[2];
-  u32 * FASTENT_RESTRICT b3 = st->bank[3];
+  u64 * FASTENT_RESTRICT b0 = st->bank[0];
+  u64 * FASTENT_RESTRICT b1 = st->bank[1];
+  u64 * FASTENT_RESTRICT b2 = st->bank[2];
+  u64 * FASTENT_RESTRICT b3 = st->bank[3];
 
   i32 mc_pos     = st->mc_pos;
   u8  m0 = st->mc_buf[0], m1 = st->mc_buf[1], m2 = st->mc_buf[2];
